python/task1/requirements.txt
argparse==1.4.0
requests==2.18.4
py-cpuinfo==9.0.0
